Move and simplify server package exports
